chore: rename workflow
