+++ game-rules-node/src/workflows/index.ts
@@ -1,1 +1,1 @@
-export { workflow as GameRulesWorkflow } from './workflow'
+export { workflow as GameRulesWorkflow } from './gameRules'
